refactor(ivy): tidy up code
