+++ packages/compiler-cli/src/ngcc/src/rendering/renderer.ts
@@ -12,19 +12,50 @@ import {AnalyzedClass, AnalyzedFile} from '../analyzer';
 import { Decorator } from '../../../ngtsc/host';
 import {ImportManager, translateStatement} from '../../../ngtsc/transform/src/translator';
 
-
+/**
+ * The results of rendering an analyzed file.
+ */
 export interface RenderResult {
+  /**
+   * The file that has been rendered.
+   */
   file: AnalyzedFile;
+  /**
+   * The rendered source file.
+   */
   source: FileInfo;
+  /**
+   * The rendered source map file.
+   */
   map: FileInfo;
 }
 
+/**
+ * Information about a file that has been rendered.
+ */
 export interface FileInfo {
+  /**
+   * Path to where the file should be written.
+   */
   path: string;
+  /**
+   * The contents of the file to be be written.
+   */
   contents: string;
 }
 
+/**
+ * A base-class for rendering an `AnalyzedClass`.
+ * Package formats have output files that must be rendered differently,
+ * Concrete sub-classes must implement the `addImports`, `addDefinitions` and
+ * `removeDecorators` abstract methods.
+ */
 export abstract class Renderer {
+  /**
+   * Render the source code and source-map for an Analyzed file.
+   * @param file The analyzed file to render.
+   * @param targetPath The absolute path where the rendered file will be written.
+   */
   renderFile(file: AnalyzedFile, targetPath: string): RenderResult {
     const importManager = new ImportManager(false);
 
@@ -38,13 +69,14 @@ export abstract class Renderer {
     });
 
     this.addImports(output, importManager.getAllImports(file.sourceFile.fileName, null));
+    // QUESTION: do we need to remove contructor param metadata and property decorators?
     this.removeDecorators(output, decoratorsToRemove);
 
     const mapPath = `${targetPath}.map`;
     const map = output.generateMap({
       source: file.sourceFile.fileName,
       file: mapPath,
-      // includeContent: true
+      // includeContent: true  // TODO: do we need to include the source?
     });
 
     return {
@@ -54,9 +86,9 @@ export abstract class Renderer {
     };
   }
 
-  abstract addImports(output: MagicString, imports: { name: string, as: string }[]): void;
-  abstract addDefinitions(output: MagicString, analyzedClass: AnalyzedClass, definitions: string): void;
-  abstract removeDecorators(output: MagicString, decoratorsToRemove: Map<ts.Node, ts.Node[]>): void;
+  protected abstract addImports(output: MagicString, imports: { name: string, as: string }[]): void;
+  protected abstract addDefinitions(output: MagicString, analyzedClass: AnalyzedClass, definitions: string): void;
+  protected abstract removeDecorators(output: MagicString, decoratorsToRemove: Map<ts.Node, ts.Node[]>): void;
 
   // Add the decorator nodes that are to be removed to a map
   // So that we can tell if we should remove the entire decorator property
