+++ packages/compiler-cli/src/ngcc/src/transform/package_transformer.ts
